Fix invalid cast in ImmutableInstruction20t

+++ dexlib2/src/main/java/org/jf/dexlib2/immutable/instruction/ImmutableInstruction20t.java
@@ -56,7 +56,7 @@ public class ImmutableInstruction20t extends ImmutableInstruction implements Ins
         }
         return new ImmutableInstruction20t(
                 instruction.getOpcode(),
-                (byte)instruction.getCodeOffset());
+                instruction.getCodeOffset());
     }
 
     @Override public int getCodeOffset() { return codeOffset; }
